Fix Converter, Add Test Program
Fixed bug in converter where delays without notes would be ignored. It's more of a Band-Aid though. Created a test playing software that will play basic files so you can get an idea of the melody. This brought to my attention that there still seems to be a bug in the conversion software that needs to be located and fixed.